Move ByteBuffer extension to separate file
