+++ Sources/Nats/NatsProto.swift
@@ -251,78 +251,6 @@ struct ServerInfo: Codable, Equatable {
     }
 }
 
-extension ByteBuffer {
-    mutating func writeClientOp(_ op: ClientOp) {
-        switch op {
-        case .publish((let subject, let reply, let payload, let headers)):
-            if let payload = payload {
-                self.reserveCapacity(
-                    minimumWritableBytes: payload.count + subject.utf8.count
-                        + NatsOperation.publish.rawValue.count + 12)
-                if headers != nil {
-                    self.writeBytes(NatsOperation.hpublish.rawBytes)
-                } else {
-                    self.writeBytes(NatsOperation.publish.rawBytes)
-                }
-                self.writeString(" ")
-                self.writeString(subject)
-                self.writeString(" ")
-                if let reply = reply {
-                    self.writeString("\(reply) ")
-                }
-                if let headers = headers {
-                    let headers = headers.toBytes()
-                    let totalLen = headers.count + payload.count
-                    let headersLen = headers.count
-                    self.writeString("\(headersLen) \(totalLen)\r\n")
-                    self.writeData(headers)
-                } else {
-                    self.writeString("\(payload.count)\r\n")
-                }
-                self.writeData(payload)
-                self.writeString("\r\n")
-            } else {
-                self.reserveCapacity(
-                    minimumWritableBytes: subject.utf8.count + NatsOperation.publish.rawValue.count
-                        + 12)
-                self.writeBytes(NatsOperation.publish.rawBytes)
-                self.writeString(" ")
-                self.writeString(subject)
-                if let reply = reply {
-                    self.writeString("\(reply) ")
-                }
-                self.writeString("\r\n")
-            }
-
-        case .subscribe((let sid, let subject, let queue)):
-            if let queue {
-                self.writeString(
-                    "\(NatsOperation.subscribe.rawValue) \(subject) \(queue) \(sid)\r\n")
-            } else {
-                self.writeString("\(NatsOperation.subscribe.rawValue) \(subject) \(sid)\r\n")
-            }
-
-        case .unsubscribe((let sid, let max)):
-            if let max {
-                self.writeString("\(NatsOperation.unsubscribe.rawValue) \(sid) \(max)\r\n")
-            } else {
-                self.writeString("\(NatsOperation.unsubscribe.rawValue) \(sid)\r\n")
-            }
-        case .connect(let info):
-            // This encode can't actually fail
-            let json = try! JSONEncoder().encode(info)
-            self.reserveCapacity(minimumWritableBytes: json.count + 5)
-            self.writeString("\(NatsOperation.connect.rawValue) ")
-            self.writeData(json)
-            self.writeString("\r\n")
-        case .ping:
-            self.writeString("\(NatsOperation.ping.rawValue)\r\n")
-        case .pong:
-            self.writeString("\(NatsOperation.pong.rawValue)\r\n")
-        }
-    }
-}
-
 enum ClientOp {
     case publish((subject: String, reply: String?, payload: Data?, headers: NatsHeaderMap?))
     case subscribe((sid: UInt64, subject: String, queue: String?))
